Refactor web live smoke into fixture-backed flow modules
